Kings Apart check and fixed console board formatting
[git-p4: depot-paths = "//depot/C35/Students/cwsullivan/SD/": change = 463225]

+++ NetChess3D/Code/Game/ChessPiece.cpp
@@ -27,7 +27,7 @@ void ChessPiece::Update()
 {
 	if ( m_moveTimer && !m_moveTimer->DecrementPeriodIfElapsed() )
 	{
-		float t = m_moveTimer->GetElapsedFraction();
+		float t = (float)m_moveTimer->GetElapsedFraction();
 		if ( m_moveStyle == MoveStyle::Slide )
 		{
 			float eased = SmoothStep5( t );
